fix(engine): carry AbiWord-width vm words + restore SV_Trace NULL mins/maxs guard at the syscall arms
Two live-bot findings from the first bots-enabled boot (bot_minplayers;
neither path is reachable by the referee's botless scenarios):

VM_Call/VM_ExplicitArgPtr/BotVMShift kept Raven's 32-bit-era int words,
truncating pointer-carrying values on LP64: the GAME_NAV_* vec3 args
(SIGSEGV on the first bot's nav query — stack pointer lost its top
bits), the GAME_CLIENT_CONNECT denied-string return, and the gWPArray
waypoint pointers through BotVMShift. Words are isize end-to-end now;
call sites widened mechanically.

Raven's SV_Trace substitutes vec3_origin for NULL mins/maxs
(sv_world.cpp:810); the by-value port made that branch unreachable
inside SV_Trace but the game module still delivers NULL words (bot-AI
trap_Trace in WPOrgVisible). The substitution moves to the G_TRACE/
G_G2TRACE/G_TRACECAPSULE arms (vma_vec3_or_origin); PORT-NOTE updated.

346 tests and all six referee scenarios byte-identical.

+++ crates/mp/engine/server/src/sv_bot.rs
@@ -367,7 +367,7 @@ pub fn SV_BotWaypointReception(sv: &mut Server, wpnum: c_int, wps: *mut *mut wpo
     let mut i: c_int = 0;
     while i < sv.bot.gWPNum {
         let wp = unsafe { *wps.offset(i as isize) };
-        sv.bot.gWPArray[i as usize] = BotVMShift(sv.gvm, wp as usize as c_int) as *mut wpobject_t;
+        sv.bot.gWPArray[i as usize] = BotVMShift(sv.gvm, wp as isize) as *mut wpobject_t;
         i += 1;
     }
 }
@@ -542,7 +542,7 @@ pub fn SV_BotFrame(common: &mut Common, sv: &mut Server, time: c_int) {
         common,
         sv.gvm,
         MpGameExport::BOTAI_START_FRAME as c_int,
-        &[time],
+        &[time as isize],
     );
 }
